Add operate function tests
